backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Google Gemini AI
google-generativeai==0.3.2

# HTTP Requests
aiohttp==3.9.1
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# CORS
python-jose[cryptography]==3.3.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
